requirements.txt
absl-py==2.2.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
grpcio==1.71.0
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
protobuf==6.30.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2024.8.30
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
